Introduce Plines::Pipeline module.
Rather than having one global pipeline (which restricts the usefulness of plines, as it could only be used by apps that have one processing pipeline), it's better to support multiple pipelines.  In this commit I've refactored so that there's no longer global pipeline state held on the Plines module.  Instead, you create a pipeline by defining a module and extending it with the Plines::Pipeline module.  All steps created within that module will be part of the pipeline automatically.