Use timecop to fix flickering acceptance test.
One of the acceptance tests was flickering based on when the ruby GC
kicked in -- by kicking in at the right moment, enough time would
pass that a timeout would happen before it was meant to. This was
only a consequence of choosing very short timeouts because we were
using `sleep` to wait for the timeouts.  It's better to use Timecop
to ensure that GC pauses don't cause the timeouts to happen at unintended
times