Verify that the jobs we meant to cancel were cancelled.
Not sure how this happened, but we've seen cases where the
job batch's meta indicates it was cancelled but the jobs
still remained. Might be a bug in Qless.

Regardless, we only want to consider the job batch as
cancelled if all jobs actually got cancelled (or were
already completed).