Store some metadata indicating when the job batch creation completed.
In prod we’re seeing some puzzling cases where there are
jobs of a batch that were enqueued after the job batch
was cancelled. We have guards against that case but it’s
still having problems and this extra bit of metadata
will help us better understand what’s going on.