Address cancellation timeouts.
For large batches it is very painful to have a cancellation
timeout, because it spends a long time preparing for it
(doing a topological sort of all jobs), which forces us
to re-do all that work when we retry. Better to have some
simple retries in place for the actual `bulk_cancel` call.